v1.7.7
Changed remindNpc logic again. 
Now we always remind in text. 
Check if beeps are enabled and remind with a beep if so.
Check if popups are enabled and remind with a popup if it is closed.
Fixed the issue where it would keep reminding you with beeps even when there was nothing to remind you about. 
Added a toggle for Show Aggro.
/am aggro will toggle on or off.
there is a toggle button on the search window as well.